🚚fix(GlobalStyle): Move GlobalStyle to styles
- 컴포넌트 폴더에 속해있던 GlobalStyle.jsx 파일을 styles 폴더로 이동한다.

by @Jeremy-Kr
