+++ src/components/index.js
@@ -1,6 +1,6 @@
 import { Footer, Header, Layout } from "./Layouts";
 import { TodoContainer, TodoInput, TodoItem, TodoList, Todo } from "./TodoList";
-import GlobalStyle from "./GlobalStyle";
+import GlobalStyle from "../styles/GlobalStyle";
 import TodoDetail from "./TodoDetail";
 
 export {
